refactor(workers): keep artifact verification within package gates
